requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
backports.zoneinfo==0.2.1
certifi==2023.7.22
charset-normalizer==3.2.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.4
django-filter==23.2
django-redis==5.3.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
idna==3.4
importlib-metadata==6.8.0
itypes==1.2.0
Jinja2==3.1.2
Markdown==3.4.4
MarkupSafe==2.1.3
openapi-codec==1.3.2
pytz==2023.3
PyYAML==6.0.1
redis==4.6.0
requests==2.31.0
simplejson==3.19.1
sqlparse==0.4.4
typing-extensions==4.7.1
uritemplate==4.1.1
urllib3==2.0.4
zipp==3.16.2
